rename deny-scene-button
